requirements.txt
jinja2
aiofiles
uvicorn
fastapi
pyyaml
